infinity

+++ src/cascadia/TerminalApp/TabHeaderControl.h
@@ -23,7 +23,7 @@ namespace winrt::TerminalApp::implementation
         WINRT_CALLBACK(PropertyChanged, Windows::UI::Xaml::Data::PropertyChangedEventHandler);
         OBSERVABLE_GETSET_PROPERTY(winrt::hstring, Title, _PropertyChangedHandlers);
         OBSERVABLE_GETSET_PROPERTY(bool, IsPaneZoomed, _PropertyChangedHandlers);
-        OBSERVABLE_GETSET_PROPERTY(uint32_t, RenamerMaxWidth, _PropertyChangedHandlers);
+        OBSERVABLE_GETSET_PROPERTY(double, RenamerMaxWidth, _PropertyChangedHandlers);
 
     private:
         bool _receivedKeyDown{ false };
